cleanup(booking): moved bookingInformation to models, removed comments
