+++ src/Models/Booking/BookingInformation.cs
@@ -1,0 +1,16 @@
+using System;
+using System.Collections.Generic;
+using StockportGovUK.NetStandard.Models.Booking.Response;
+
+namespace form_builder.Models.Booking
+{
+        public class BookingInformation
+        {
+            public DateTime CurrentSearchedMonth { get; set; }
+            public DateTime FirstAvailableMonth { get; set; }
+            public List<AvailabilityDayResponse> Appointments { get; set; }
+            public bool IsFullDayAppointment { get; set; }
+            public DateTime AppointmentStartTime { get; set; }
+            public DateTime AppointmentEndTime { get; set; }
+        }
+}
